Refactor : Notice 페이지  error, loading,layout 컴포넌트 구조 개선 & Board 페이지 폴더 구조 개선
