+++ src/app/(home)/(board)/board-write/components/board-input.tsx
