+++ src/app/(home)/(board)/board-write/components/helper-board-form.tsx
@@ -4,15 +4,15 @@ import { useRouter } from 'next/navigation';
 import { postBoardData } from '@/actions/common';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
-import BoardInput from '../board-input';
+import BoardInput from './board-input';
 import Button from '@/components/ui/button';
-import BoardMesoInput from '../board-meso-input';
+import BoardMesoInput from './board-meso-input';
 
 import { IHelperBoardPost } from '@/types';
-import BoardRadio from '../board-radio';
-import BoardSelect from '../board-select';
+import BoardRadio from './board-radio';
+import BoardSelect from './board-select';
 import usePost from '@/hooks/use-post';
-import BoardTimeInput from '../board-time-input';
+import BoardTimeInput from './board-time-input';
 import toast from 'react-hot-toast';
 import Icon from '@/components/ui/icon';
