+++ src/app/(home)/(board)/board-write/components/wood-cutter-board-form.tsx
@@ -1,28 +1,27 @@
 'use client';
 
 import usePost from '@/hooks/use-post';
-import { IPartyBoardPost } from '@/types';
+import { IWoodCutterBoardPost } from '@/types';
 import { useRouter } from 'next/navigation';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import BoardInput from '../board-input';
-import BoardTimeInput from '../board-time-input';
-import BoardRadio from '../board-radio';
+import BoardMesoInput from './board-meso-input';
+import BoardInput from './board-input';
+import BoardRadio from './board-radio';
+import BoardSelect from './board-select';
+import BoardTimeInput from './board-time-input';
 import Button from '@/components/ui/button';
-import BoardFloorInput from './board-floor-input';
-import Label from '@/components/ui/label';
-import BoardFloors from './board-floors';
 
-interface IPartyBoardFormProps {}
+interface IWoodCutterBoardFormProps {}
 
-const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
-  const { control, handleSubmit } = useForm<IPartyBoardPost>();
+const WoodCutterBoardForm: React.FunctionComponent<IWoodCutterBoardFormProps> = (props) => {
+  const { control, handleSubmit, watch } = useForm<IWoodCutterBoardPost>();
   const { isLoading, setIsLoading, isError, setIsError } = usePost();
   const router = useRouter();
 
-  const onSubmit: SubmitHandler<IPartyBoardPost> = async (data) => {
+  const onSubmit: SubmitHandler<IWoodCutterBoardPost> = async (data) => {
     setIsLoading(true);
     try {
-      if (data.progress_time === '협의 가능') {
+      if (data.meso === '협의 가능') {
         const newData = { ...data, meso: null };
         // TODO : fetch New Data
         setIsLoading(false);
@@ -39,6 +38,8 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
     }
   };
 
+  const jobWatch = watch('main_job') || null;
+
   return (
     <>
       {isError ? (
@@ -48,6 +49,19 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
       ) : null}
       <form onSubmit={handleSubmit(onSubmit)} className='grid grid-cols-2 gap-x-8'>
         <div className='flex flex-col gap-y-8'>
+          <BoardMesoInput
+            name='meso'
+            control={control}
+            disabled={isLoading}
+            rules={{
+              required: '메소는 필수로 입력해야 합니다.',
+              pattern: {
+                value: /^[0-9]*$|^협의 가능$/,
+                message: '숫자만 입력 가능합니다.'
+              }
+            }}
+          />
+
           <BoardInput
             control={control}
             name='title'
@@ -61,7 +75,7 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
                 value: 5
               },
               maxLength: {
-                message: '50글자 이상은 입력이 불가합니다.',
+                message: '30글자 이상은 입력이 불가합니다.',
                 value: 30
               }
             }}
@@ -90,19 +104,44 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
             control={control}
             name='hunting_ground'
             label='사냥터'
-            placeholder='사냥터를 입력해 주세요'
             disabled={isLoading}
+            placeholder='사냥터를 입력해 주세요'
             rules={{
-              required: '사냥터는 필수로 입력해야 합니다.',
+              required: '제목은 필수로 입력해야 합니다.',
               minLength: {
-                message: '최소 2글자 이상 입력해야 합니다.',
-                value: 2
+                message: '최소 5글자 이상 입력해야 합니다.',
+                value: 5
               },
               maxLength: {
                 message: '20글자 이상은 입력이 불가합니다.',
-                value: 12
+                value: 20
               }
             }}
+          />
+        </div>
+        <div className='flex flex-col gap-y-8'>
+          <BoardRadio
+            control={control}
+            name='main_job'
+            label='직업'
+            disabled={isLoading}
+            rules={{ required: '직업을 선택해 주세요' }}
+            options={[
+              { id: '전사', label: '전사', value: '전사' },
+              { id: '마법사', label: '마법사', value: '마법사' },
+              { id: '도적', label: '도적', value: '도적' },
+              { id: '궁수', label: '궁수', value: '궁수' }
+            ]}
+          />
+
+          <BoardSelect
+            control={control}
+            jobWatch={jobWatch}
+            label='서브 직업'
+            name='sub_job'
+            disabled={isLoading}
+            placeholder='서브 직업을 선택해 주세요'
+            rules={{ required: '서브 직업을 선택해 주세요' }}
           />
 
           <BoardTimeInput
@@ -127,24 +166,11 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
             }}
           />
 
-          <BoardRadio
-            control={control}
-            name='main_job'
-            label='주차'
-            disabled={isLoading}
-            rules={{ required: '주차 여부를 선택해 주세요' }}
-            options={[
-              { id: '주차 가능', label: '주차 가능', value: 'true' },
-              { id: '주차 불가', label: '주차 불가', value: '' }
-            ]}
-          />
-        </div>
-        <div className='flex flex-col gap-y-8'>
           <BoardInput
             control={control}
-            name='recruit_people_count'
-            label='모집 인원'
-            placeholder='모집 인원 수'
+            name='level'
+            label='레벨'
+            placeholder='레벨을 입력해 주세요'
             disabled={isLoading}
             type='number'
             rules={{
@@ -163,12 +189,12 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
                 value: 10
               }
             }}
-            icon={<span className=' items-center text-[15px] font-semibold text-black'>명</span>}
+            icon={<span className=' items-center text-[15px] font-semibold text-black'>LV</span>}
           />
-          <BoardFloors control={control} />
         </div>
+
         <div></div>
-        <div className='mt-4'>
+        <div className='mt-12'>
           <Button size='wide' color='main' disabled={isLoading}>
             등록하기
           </Button>
@@ -178,4 +204,4 @@ const PartyBoardForm: React.FunctionComponent<IPartyBoardFormProps> = ({}) => {
   );
 };
 
-export default PartyBoardForm;
+export default WoodCutterBoardForm;
