+++ src/app/(home)/(board)/components/ui/board-inner.tsx
