+++ src/app/(home)/(board)/components/ui/board-result.tsx
