+++ src/app/(home)/(board)/components/ui/post-card.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from 'react';
 
-import Icon from '../../../../components/ui/icon';
-import Badge from '../../../../components/ui/badge';
-import Button from '../../../../components/ui/button';
+import Icon from '../../../../../components/ui/icon';
+import Badge from '../../../../../components/ui/badge';
+import Button from '../../../../../components/ui/button';
 import InlineProfile from '@/components/ui/inline-profile';
 
 import dayjs from 'dayjs';
